requirements.txt
bcrypt>=4.0.1
email-validator>=2.0.0
flask>=2.3.0
flask-login>=0.6.2
flask-sqlalchemy>=3.0.0
flask-wtf>=1.1.1
gunicorn>=20.1.0
psycopg2-binary>=2.9.5
pycryptodome>=3.18.0
pyotp>=2.8.0
python-dotenv>=1.0.0
python-magic>=0.4.27
qrcode>=7.4.2
sqlalchemy>=2.0.0
werkzeug>=2.3.0
wtforms>=3.0.0